[refactor] pages 폴더로 정리
