+++ src/App.jsx
@@ -3,29 +3,31 @@ import "./App.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import NavigationBarLayout from "./layouts/NavigationBarLayout";
 import { MobileResponsiveLayout } from "./layouts/MobileResponsiveLayout";
-import Signup, { SIGNUP_PAGE_PATH } from "./pages/Signup";
-import Signin, { SIGNIN_PAGE_PATH } from "./pages/Signin";
+import Signup, { SIGNUP_PAGE_PATH } from "./pages/auths/Signup";
+import Signin, { SIGNIN_PAGE_PATH } from "./pages/auths/Signin";
 import DiagnosisGuide, {
   DIAGNOSIS_GUIDE_PAGE_PATH,
 } from "./pages/dementiaDiagnosis/DiagnosisGuide";
 import Diagnosis, {
   DIAGNOSIS_PAGE_PATH,
 } from "./pages/dementiaDiagnosis/Diagnosis";
-import DiaryDraw, { DIARY_DRAW_PAGE_PATH } from "./pages/DiaryDraw";
+import DiaryDraw, { DIARY_DRAW_PAGE_PATH } from "./pages/draws/DiaryDraw";
 import Home, { HOME_PAGE_PATH } from "./pages/Home";
-import Diary, { DIARY_PAGE_PATH } from "./pages/Diary";
+import Diary, { DIARY_PAGE_PATH } from "./pages/diarys/Diary";
 import HeaderBarLayout from "./layouts/HeaderBarLayout";
 import PaperBackgroundWrapper from "./components/diary/PaperBackgroundWrapper";
 import React from "react";
 import MyPage, { MY_PAGE_PATH } from "./pages/MyPage";
-import DiaryRecall, { DIARY_RECALL_PAGE_PATH } from "./pages/DiaryRecall";
+import DiaryRecall, {
+  DIARY_RECALL_PAGE_PATH,
+} from "./pages/diarys/DiaryRecall";
 import NoDiaryRecall, {
   NO_DIARY_RECALL_PAGE_PATH,
-} from "./pages/NoDiaryRecall";
+} from "./pages/diarys/NoDiaryRecall";
 import RecentDiaries, {
   RECENT_DIARIES_PAGE_PATH,
 } from "./pages/mypage/RecentDiaries";
-import HelpWithAi, { HELP_WITH_AI_PATH } from "./pages/HelpWithAi";
+import HelpWithAi, { HELP_WITH_AI_PATH } from "./pages/draws/HelpWithAi";
 import ControlPhotoOpacity, {
   CONTROL_PHOTO_OPACITY_PAGE_PATH,
 } from "./pages/draws/ControlPhotoOpacity";
