+++ src/pages/auths/Signup.jsx
@@ -1,10 +1,10 @@
-import HeaderBar from "../components/HeaderBar";
-import SignupSlider from "../components/signup/SignupSlider";
-import { useInput } from "../hooks/useInput";
-import InputStep1 from "../components/signup/InputStep1";
+import HeaderBar from "../../components/HeaderBar";
+import SignupSlider from "../../components/signup/SignupSlider";
+import { useInput } from "../../hooks/useInput";
+import InputStep1 from "../../components/signup/InputStep1";
 import React, { useState } from "react";
-import InputStep2 from "../components/signup/InputStep2";
-import SigupStep from "../components/signup/SignupStep";
+import InputStep2 from "../../components/signup/InputStep2";
+import SigupStep from "../../components/signup/SignupStep";
 
 export const SIGNUP_PAGE_PATH = "/signup";
